[lucicfg] Add support for luci-config, logdog and buildbucket ACLs.
ACLs can be defined either on a project or on a bucket levels. ACLs defined on
the project level are inherited by all buckets.

Additionally, some ACLs (such as PROJECT_CONFIGS_READER, LOGDOG_*) can be
defined only on the project level, since LUCI doesn't support defining them on
a bucket level yet or they make no sense when defined on a bucket level.

R=nodir@chromium.org, iannucci@chromium.org
BUG=833946

Change-Id: Ib6175cb2cdc83aae7300c5e2a731882f6ecc4430