[milo] Only show timeline for LUCI builds.
As part of the build.proto refactor, the timeline code will be increasingly
specific to build.proto.  To avoid maintaining two codepaths for timeline rendering,
we only want to support one, which is to show the timeline for LUCI builds only.

Also:
split tabs related CSS from timeline.css out to tabs.css

Change-Id: Ia3bd84e52a10c3972846c90a4917a394ffcab362