Add bug links functionality for build pages
Add the ability to put links on the build pages intended
to allow prefilling of bug entry forms. Configuration protos
are changed to allow the entering of an optional BugTemplate
for a project. The template features summary and description
placeholders that get interpolated by the context of the
build. If all the interpolation placeholders for a particular
template are satisfied that link is rendered on the page.

The template syntax follows that of the built in go
templates. That is {{.placheholder}} style. The placeholders
that are allowed are documented in the proto definition file.
Additional placeholders are added easily if the data needed
can be obtained from either query parameters or the MiloBuild
data instance.

Bug: 903385
Change-Id: I9f551772d14d0b90f1adcf87bfa1d029446486d3