[GCE] Reset hostname if instance creation fails
Since request ID is derived from the hostname, the service will continue
issuing the same idempotent request if instance creation fails. We
actually want to attempt the creation again if this happens. The
simplest way to do that is to reset the hostname and allow a new one to
be generated.

Bug: 897355
Change-Id: I13e2d2fed5b22bbda4c8ed76f2502917bdaa34d7
Reviewed-on: https://chromium-review.googlesource.com/c/1375312
Commit-Queue: smut <smut@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>