[milo] Introduce build.html based off of build.proto
This adds a build.html that takes in a "BuildPage" struct,
which wraps a buildbucket build.proto struct.

This implementation is incomplete, but is accessible behind a hidden ?v2=1 flag.

This also breaks Timeline, which previously depended on MiloBuild,
but now depends on build.proto.

A sample live build was captured from today as testdata.

Bug: 850113
Change-Id: I417a19a65d4a1f7c2a942d648a12b57f07c4bde4
Reviewed-on: https://chromium-review.googlesource.com/c/1375309
Commit-Queue: Ryan Tseng <hinoka@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>