[lucicfg] Make graph.key objects introspectable through Starlark.
They now have attributes 'container', 'kind', 'id'. This would become useful
for implementing rules that behave differently depending on a kind of an object
they are related to.

For the key [(k1, i1), (k2, i2), (k3, i3)]:
  'container' is key [(k1, i1), (k2, i2)]
  'kind' is string 'k3'
  'id' is string 'i3'

Term "container" was chosen to avoid confusion with term "parent", already used
in the graph API to denote a different sort of relation.

R=nodir@chromium.org, tandrii@chromium.org
BUG=833946

Change-Id: I16ea3709b6a2adbff9a9d7f810d1080cd5c48f24