Add {{.MiloBuildUrl}} as build page feedback link interpolation placeholder
Add the placeholder {{.MiloBuildUrl}} as a placeholder that can be used in
BugTemplates for formulating the feedback URL. This allows the user to easily
insert a URL back to the build page in the description of the bug created
by the feedback link.

This feature is added to the legacy page - presumably when the new build page
refactoring is more complete and the dust settles there this functionality will
likewise be moved to that page.

Bug: 903385
Change-Id: I2310608df3cac9755747cb04529021fd5b7be300