Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> Bikeshedding, but i like {{.URL}} or {{.SelfURL}} better.

Actually, I've been thinking that the other direction is better. That is, back to {{.MiloBuildUrl}} rather than the {{.MiloUrl}} that I currently have in this CL.
Reason being that someone may want to formulate a link that points back to the
failing build {{.MiloBuildUrl}} and all recent builds for that builder, which could
then be done as {{.MiloBuilderUrl}}. For example someone may wish to put together
a bug template link that links both to the failing build:

{{.MiloBuildUrl}}
https://ci.chromium.org/p/chromium/builders/luci.chromium.ci/Mac10.12%20Tests/18004

and recent runs for that builder:

{{.MiloBuilderUrl}}
https://ci.chromium.org/p/chromium/builders/luci.chromium.ci/Mac10.12%20Tests

This seems like a better way to go to me. Thoughts?

Patch-set: 5
Reviewer: Gerrit User 1297361 <1297361@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>