Add {{.MiloBuilderUrl}} as build page bug template placeholder
Add the placeholder {{.MiloBuilderUrl}} as a placeholder that can be used
in BugTemplates for formulating the feedback URL on the build page. This
allows a user to easily insert a URL back to the builder page in the
description of the bug created by the feedback link.

This feature is added to the legacy page - presumably when the new build page
refactoring is more complete and the dust settles there this functionality will
likewise be moved to that page.

Bug: 903385
Change-Id: Ia29515b7c225acd7a49e92b1176b3c450c453077