Update patch set 1
Patch Set 1:

PTAL and comment of whether this is understandable at all.

I'm still experimenting with this (will try to add triggers in the next CL and see how it looks). Will add tests once I'm confident this is the way to go.

core.builder_group is not absolutely necessary, but:
  1. I wanted to add to see what problems it may introduce (and I did see them).
  2. I suspect it may be useful later, e.g. I saw early pre-release design doc from Sergey about chromium/src configs and it had builder groups concept.

I'm a bit sad about the complexity :( I'm confident the graph model is powerful enough to support all our use cases, but the implementation may look unnatural... Most of the complexity comes from requirement (1) (allowing omitting bucket name).

Note that the public interface is still very simple and natural. The added complexity is internal. (Due to my insistence on using order-independent declarations that build a graph, but are not allowed to examine it).

Patch-set: 1
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false