[lucicfg] Initial empty definition of core.builder rule.
There an aspect that makes the implementation somewhat tricky: we want to allow
to refer to builders using their names alone (without specifying the bucket
name), if this is unambiguous. For cases when there are more than one builder
with a given name, the ambiguity can be resolved by using the full builder name
(e.g. "ci/linux builder").

This is solved by introduction of intermediate 'builder_ref' node:
  * Nodes that refer to builders have builder_refs as children.
  * A builder_ref has a builder as a child.
  * A single builder has two parent builder_ref's: one has bucket-scoped
    name (e.g. "ci/linux builder") and another is global (e.g. "linux builder").
  * Ambiguities manifest as builder_ref nodes with more than one child.

There's an internal API to follow such refs.

R=nodir@chromium.org, tandrii@chromium.org
BUG=833946

Change-Id: I1c126dd3f387a5915f4c49079b05acb1d95b45c8