Update patch set 11
Patch Set 11:

(2 comments)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > what is builder group? is this the same concept as builder groups in milo? The latter must have a category for each builder ref. They are used to build console UI. How this will be handled?
> 
> Where are these Milo builder groups defined and what are their relation to milo Consoles?

Due to the mess with milo concepts that I mentioned before, today a builder group is defined by defining a console. It should have been fixed, but it was never done crbug.com/788995

> There are no mentions of them in Milo's project.proto [1]. 

because it is a mess

> I see category in Builder message though. And it looks to me like category is relevant only to consoles (since Milo's builder view ignores it). 

correct

> What fields of Console message are relevant to this more general builder group concept?

none

"none" doesn't mean that we can just ignore the console-case though. This config language should be user-friendly, so i think we can define a per-builder console-specific field in a builder group. My point primarily is that [builder] might be insufficient, and may have to do [builderEntry] where entry may have extra fields besides a reference to a builder. But I don't have a strong opinion _how_ to solve this problem, but we have to take it into account

> 
> [1] https://chromium.googlesource.com/infra/luci/luci-go/+/master/milo/api/config/project.proto

Patch-set: 11