[lucicfg] Add validate.int(), validate.bool() and validate.str_dict().
validate.int will be used to validate e.g. swarming priority
validate.bool will be used to validate Toggle-typed buildbucket fields.
validate.str_dict will be used to validate e.g. dimensions and properties.

Adjust the interface of the validators slightly to allow None to be used as
a value for required=False fields. It is useful to propagate None as "hasn't
been specified" flag (or equivalently to represent Toggle.UNSET), instead of
replacing it with some default value right away.

R=tandrii@chromium.org, nodir@chromium.org
BUG=833946

Change-Id: Ifbdc9a740972803e7f6d4bb707b07bcd418fb74b