[lucicfg] docgen: Add docstring parser.
It parses a docstring into a main body, a list of arguments and a list of
additional sections, like "Returns". All extracted strings are properly
deindented and cleanup, ready to be inserted into e.g. markdown.

This is surprisingly non-trivial, complicated by careful handling of
indentation, since it is syntactically significant in this case.

R=tandrii@chromium.org
BUG=833946

Change-Id: I38c1f99cb71331e2f6617196415b4e914e15614e