[lucicfg] docgen: Add AST parser for the documentation generator.
It takes a Starlark's AST and reduces it only to structures important for the
documentation generation purposes, preserving their docstring and comments.

It works on a single file only. It will be used by a higher level parser that
understands relations between modules in a package, e.g. if a symbol is defined
in an internal module, but then publicly exposed through an alias in another
module, the documentation generator will be able to trace the references back
to the original definition, to grab a documentation from it.

R=tandrii@chromium.org
BUG=833946

Change-Id: Ic5ae1ac25978cb3964f20d3ae43bc3265fc7078c