[lucicfg] docgen: Add the actual generator.
It is a text generator based on text/template that has access to functions that
return structured information about symbols defined in the starlark code. This
allows templates to insert docs extracted from the source code at appropriate
places.

This CL has only very basic implementation, as a proof of concept. It needs
a lot more polish before it generates acceptably looking and useful docs.

R=tandrii@chromium.org
BUG=833946

Change-Id: I849ace37b62782f25ac2603d968016a709728f5a