Update patch set 1
Patch Set 1:

check out this gross hack

In swarming, the shard index and total shard count is a property of the task.  In tumble, the shard index is a property of the task, but total shard count is a property of the global config.  That's why this looks like this :(

Patch-set: 1