[lucicfg] Create Chrome OS source tree test config proto.
This proto is intended as the output format for the Chrome OS source
tree testing config, as generated by lucicfg's Starlark interpretter.

The proto unfortunately needs to live in the lucicfg repo so that it can
be compiled into the binary. Rumor has it that the Go proto package will
eventually allow runtime inclusion of external protos.

BUG=chromium:922998
TEST=Built lucicfg and ran it over the .star file that uses this new proto

Change-Id: Id1add14ee8d993329b0485c59a0f126fb1d22eb9