Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

this code looks correct, but i don't see the big picture. +Andrii. In particular, instead of doing this dance of fetching a list of items and then making an RPC for each item (on the same server), why not add a new RPC that does all this (in one RPC)?

is this actually an RPC or it happens in the same process and we are using protobufs where plain old Go interfaces and structs would do?

Patch-set: 1
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1