Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> this code looks correct, but i don't see the big picture. +Andrii. In particular, instead of doing this dance of fetching a list of items and then making an RPC for each item (on the same server), why not add a new RPC that does all this (in one RPC)?

It could be batched, but I find it easier to read and debug if it's one at a time. This is similar to config import in the machine-db/appengine/{config,model} packages: for each configured entity we create or update it in a separate db operation, rather than doing bulk operations to create or update them all at once. Except instead of an RPC server indirection layer machine-db has some other indirection layer (the *Table structs in model).

> is this actually an RPC or it happens in the same process and we are using protobufs where plain old Go interfaces and structs would do?

It's not really an RPC; it doesn't use network. We have a local instance of the same struct which is installed as an HTTP handler and we just invoke those same functions directly. In other words, we're just doing a function call, but it's the same function call that would occur if a user issued an HTTP request to the right place.

I suppose go structs would do, but the config is imported in protobuf form and converting it to a struct would mean more places to update each time a field is added, so I keep it mostly in protobuf form throughout the app.

The RPC server's HTTP routes are not really meant to be used outside of development and debugging (and are already ACLed and documented as such), but they greatly eased that process since I didn't have to make commits to update a config each time I wanted to test something. We could even uninstall the HTTP routes in gce/appengine/frontend at this point.

Patch-set: 1
Reviewer: Gerrit User 1003211 <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0