Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> my repro was
> 1. Go https://ci.chromium.org/p/chromium/builders/ci/Linux%20Tests/75476 (LinuxTestsManySteps was based off this)
> 2. Immediately after page loads (within 500ms, this is key) click on the "timeline" tab
> 
> It's been reliably reproducible on my machine (OSX)
> 
> Double render isn't good, maybe this should test for visibility instead.

Maybe you can't see it on your machine. On my machine when that hack is in place - and I render that timeline with VERY MANY steps. I briefly see a bunch of the steps vertically down the middle with a red vertical line and then quickly it renders normally. I'll see if I can figure out how to capture a video of it if you are not seeing it on OSX.

Patch-set: 4
Reviewer: Gerrit User 1297361 <1297361@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>