[starlark] Print full stack trace when some deep load(...) fails.
This is similar to https://chromium-review.googlesource.com/c/1444569 but for
load(...).  It was harder to spot load(...) had the same problem because there's
usually not many errors in the top level of a loaded script (it is just a bunch
of function definitions).

R=tandrii@chromium.org
BUG=833946

Change-Id: I15996fc8da5da71954ec7426a0bf6ddec336ea68
Reviewed-on: https://chromium-review.googlesource.com/c/1449309
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>