[lucicfg] Make var.get() auto-initialize the variable to the supplied default.
It means once var.get() is called for the first time, all subsequent calls
will observe the exact same value (either an explicitly set one or the default).

This is a measure to reduce "statefullness" of vars, meant to outlaw the
following case:

  lib.do_something()  # does something based on lib.var default value
  lib.var.set(123)
  lib.do_something()  # suddenly does something else, since lib.var is set now

Now lib.var.set(...) call will fail with a message explaining that it was
auto-initialized as a side effect of first 'get'.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=833946

Change-Id: I0efb87cc0833d3e74ac9e19741b27f83ea166476