[isolated] Add partial support for SHA256 and SHA512
Add the needed piping of the namespace to use the right hash algorithm, it used
to be hardcoded to SHA1.

Use crypto.Hash instead of passing a string around for performance critical
functions, to reduce string processing.

Increase unit test coverage a bit. Still more to do but I didn't want to make
the CL too large.

isolated.New() still hardcode SHA1, so there's more to do but that'll be done in
a follow up.

Bug: 787113
Change-Id: I59322a4408a841ed1153b49525ac19188b5aef90
Reviewed-on: https://chromium-review.googlesource.com/c/1450656
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>