[lucicfg] Add `strutil.generate_name` helper function.
It derives a unique (within single lucicfg invocation) name based on the hash
of inputs. Will be used to derive names of some nodes from their content.

Structure of such names is internal implementation detail of lucicfg and it must
not be relied upon from external configs. It may change any time. This will be
documented in rules that use such names.

R=nodir@chromium.org
BUG=833946

Change-Id: I74a6632aca3567155f25c8b270e9e92d928c1969