[config] Stop caching var expansions in validation rules.
This is in preparation for adding ${config_service_appid} that would expand into
an app ID of a LUCI config instance configured for the app.

Turns out caching on this layer is not really necessary, since the callbacks
that provides values (for ${appid} and ${config_service_appid}) implement the
caching already.

R=nodir@chromium.org, fmatenaar@chromium.org

Change-Id: I62c1e3846ce6abd339d4dc72dd9fded682ab0884
Reviewed-on: https://chromium-review.googlesource.com/c/1467881
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Felix Matenaar <fmatenaar@chromium.org>