[GCE] Do not update VM entities with config-related changes
There are a few subtleties related to updating VMs. For example, it's
not safe to update the Swarming server once an instance exists:
1. Instance is created and connects to Swarming server A.
2. VM is updated to use Swarming server B.
3. Backend now thinks the instance belongs to server B.
The backend can no longer detect the instance's presence on its Swarming
server, or terminate it properly.

As another example, it's generally safe to update attributes, except
when one of the changing attributes is zone. Then there can be a race
condition if instance creation has started but not completed:
1. Instance creation is started using zone A.
2. VM is updated to use zone B.
3. Instance creation progress is checked in zone B.
The instance will not be found, even if it's successfully created.

Since different fields of the VM are safe to be updated at different
times, the easiest solution is to just treat these entities as
immutable. Don't allow config-related modifications to a VM at any time,
but delete the VM every time an instance is destroyed. Then config
updates will only be reflected in the VM at entity creation time, which
is necessarily when no corresponding instance exists or will exist.

Bug: 897355
Change-Id: I3a07ffedbed2738b4aeb3837f2c256c573176bd8