[buildbucket] Add Build.output.critical
A Chrome OS build will indicate if it is critical, see the definition in the proto.
All "important" builders will be critical most of the time, except periods of time
when some builders were temporarily marked as non-critical/snoozed.
The marking is Chrome OS impl detail. The contract with LUCI is a build can report its criticality.

Milo and Gerrit-Buildbucket plugin will gray out (paint with more transparent color)
non-critical builds.

Based on the field definition, ideally it should be honored by CQ.
It does not matter for Chrome OS because Test Orchestrator will interpret the value.

Additional work would be needed in recipe land to actually set this value, probably after
recipe engine switches to build.proto

The field is Trinary and not bool because otherwise criticality=false would be default.
This would be a bad default and a breaking change.

Bug: 924657
Change-Id: I0a9a259477a58ab75605a9f083116fdc5939a66c