[starlarkproto] Rename *builtinWithAttrs to *messageCtor.
builtinWithAttrs represents a callable that constructs proto messages. It will
be passed by value to `from_pbtext` and `from_jsonpb` functions to specify a
type of the message to deserialize.

This CL makes the type name more concrete and adds *MessageType to it. It will
be used in the later CLs that implement proto deserialization.

R=tandrii@chromium.org, nodir@chromium.org
BUG=833946

Change-Id: I41cc54b6b9ac4caeadfe3094c631e829a30df3d8