[GCE] Allow empty configs, refactor package, add normalization tests
A missing config is better interpreted as a valid empty config, because
if the kinds.cfg and vms.cfg files were present but completely empty,
that would constitute a valid configuration for the service.

Renames cfgs to vms for clarity, since there are different types of
configs.

Groups kinds and vms into a single Config struct rather than passing
around multiple values.

Adds test for normalize.

Bug: 897355
Change-Id: I2de486f59080bec531ed2921e80ca795da6b26f0