[luccifg] Require most arguments to be passed via keyword args.
This ensures better forward compatibility and improves readability of configs.

Exceptions to this rule are some constructors that allow one and only one
positional argument (whose meaning is clear from the name of the function):
  * acl.entry(<role>)
  * swarming.cache(<path>)
  * swarming.dimension(<value>)
  * luci.list_view_entry(<builder>)
  * luci.console_view_entry(<builder>)

This is a breaking change, bump the version number to 1.4.0.

R=tandrii@chromium.org
BUG=833946

Change-Id: Ib0944b8d0b31df60a6a888e2437158afd4d39666
Reviewed-on: https://chromium-review.googlesource.com/c/1484658
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>