[milo] Use stacked bubbles for console's expanded view
This changes the behavior of the [expand] function.
In the new expand mode, the build bubbles are able to flow vertically,
which makes more efficient use of the whitespace.

The implementation is done via DOM manipulation in javascript,
by combining all builders within a "leaf category"
(A category where all children are builders) into a single column.

Bug: 903491

Change-Id: I6435247aab5945aeb4653abfea9275ce014f643f