Update patch set 1
Patch Set 1:

PTAL

There are no tests because I consider this to be hacky temporary tool to be removed when most projects are migrated to lucicfg. It either works or it doesn't (that's the test). I'm not concerned about it rotting with time.

Patch-set: 1
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false