Add dummy BuilderConfig proto to lucicfg external protos.
Add an empty dummy BuilderConfig proto to lucicfg's external protos allow it to be generated
using lucicfg. The intention here is to get this in place to make it easier to begin iterating
on ideas for the starlark code that will generate the config.

BUG=923000

Change-Id: I735410947a15cb21157fc3139a625985765350ef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1500558
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>