Update patch set 2
Patch Set 2: Code-Review+1

Not that I'm really a reviewer for this sort of thing, but this looks right-ish :).

Does it make sense to add a metric for the time to complete the overall request as well, if we don't have one?

Patch-set: 2
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1