[logdog] Add taskqueue-backed archive task leasing endpoint
We want to use GAE TQ for:
* Pull Queues
* Configurable Retries
* Delays
* Stable API and SLO

Neither Cloud PubSub nor Cloud Tasks (Beta) fulfill all of these requirements.

But because we cannot lease GAE TQ tasks in GKE, we need to build a wrapper
around LeaseTasks and DeleteTasks.

This CL implements the minimal amount of logic for GKE to lease and delete tasks.

Bug: 923557
Change-Id: Ia5b99150f020ea3f4f89e159d6dfcfeace541e81