[auth-client] Implement fallback when AsProject fails with NotFound.
During migration to project scoped service accounts, clients will not
know whether a particular project has a scoped service account
configured or not. This implementation will allow AsProject to
perform a fallback to obtain a service token instead in case the
tokenserver indicates that the project scoped identity was not found.

R: vadimsh@chromium.org
Bug: 911811
Change-Id: Id7e1c78c4cdeb37ebfb3e05a4029ad37b9f73123