[scheduler] Failed LaunchTask after a retry should not be retried.
If a task first fails transiently (e.g. HTTP 500 from Buildbucket), but then (on
a retry) fatally (e.g. HTTP 400 from Buildbucket), it must not be retried.

There was a bug in the Scheduler where StatusRetrying state was "sticky" and
didn't reset to StatusFailed if LaunchTask fails. Fix it.

Also log "The invocation will be retried" exactly once instead of in every
Save().

R=tandrii@chromium.org

Change-Id: Ibcdbd26acb39e08d2dbd5549f03ef883621b9662