[GCE] Don't abort naming drained VMs
A named, drained VM proceeds with instance creation and manageBot takes
care of noticing the VM is drained and safely deleting it. An unnamed,
drained VM is not named, which means instance creation never proceeds,
but its deletion is never handled either because manageBot is the only
place which handles deleting drained VMs, and it only runs for created
instances.

Rather than adding a second place which handles deleting drained VMs,
simply proceed with naming and instance creation, letting manageBot take
care of this case as well instead of treating an unnamed drained VM as a
special case.

Bug: 897355
Change-Id: Ifaa3d5d5324f8af3c85016d0f632a958f29f75fa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1529959
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: smut <smut@google.com>