Add buildbucket/bootstrap package
There will be a new Go binary, replacement for kitchen, that does not know about
recipes, annotations, but will take the best parts of kitchen, which is only
auth stack. Most of the binary will be implemented as a reusable package,
this one. For now, only move auth stack here and update comments.

Bug: 939368
Change-Id: I0444165fee0f3f210ff2c60d1fd7a8c16896c371