Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (4 comments)
> 
> lgtm in general, but what immediate problem does this change solve? I think we are in a bit of turmoil with simultaneous migrations to v2 buildbucket API, lucicfg configs, and short bucket names. Can we delay deprecating "recipe {...}" field until some migrations are finished? E.g. once we use lucicfg everywhere, switching to a new proto format will be almost trivial.

This is a blocker for recipe parallelism. See the doc linked in the comment description.

v2 buildbucket API / short bucket names are orthogonal to this?

i can make it so specifying Builder.executable and Builder.properties with "$recipe_engine" (with recipe name) is identical to specifying Builder.recipe, and update lucicfg to generate those, so that you are not blocked on this.

We were thinking of using Builder.executable vs Builder.recipe as the switch between old and code path (kitchen and recipe engine guts), but now I am not sure: don't want to hunt user configs to make them switch to the new code path. I'd rather prefer to have the whitelist somewhere in a foundation-owned config.

WDYT?

Patch-set: 7
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0