[milo] Fix relative time builders view
Previously a change was added to show all builders in a project,
but didn't take into account the "groups" mode in builders view.

This changes the behavior such that swarmbucket data is used
for groups mode only if config mode didn't return anything.
This uses bucket as the implicit group.

Bug: 946294
Change-Id: I76a99fc3b5121ad1ebaa8c08e1487ae3c22da739