Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > PTAL
> > 
> > I consider making infra.git's infra/config/generated/* directory the root of the "projects/infra" config set. It means ALL config files should end up there. Also I want this directory to be 100% generated.
> > 
> > luci.emit(...) will be useful for this. I plan to use for tricium.cfg and luci-notify template files. Though I'm not 100% sure yet. Maybe having a mess with manually edited and generated files all in one directory is better?
> 
> With luci.emit() i'd still need to mess with file manually, just in a different dir. I think that's worse than modifying file in a dir with other files being generated. 

So there are 3 alternatives:

1. Have one directory with everything: *.star files, manually edited *.cfg, generated *.cfg.
2. Have two directories: one with *.star files, another with all configs (edited and generated).
3. Have two directories: one with *.star files and manually edited configs, another is completely auto-generated.

Why I don't like 1 and 2 - it is not clear from a glance what is actually editable and what's not.

Why I additionally like 3: by generating EVERYTHING we also know what to send to LUCI Config for validation (or even as a push, if we ever switch to the push model). Otherwise we'll either need to send all **/* files, or somehow tell lucicfg which files it should pick up as belonging to the config set. I actually initially started with that, and arrived to lucicfg.emit(...) as a solution to do that (you want a file to be in a config set? emit it). 

Though I do agree having two identical files is confusing (though presubmits will verify they are indeed identical). Maybe we can rename "generated" into ".generated" to hint not to mess with it manually?

> Of course, that means that ideally the dirname isn't "generated".

Do you have proposal for how to name it?

Patch-set: 1
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>