[lucicfg] Use full file path when matching against tracked_files patterns.
Otherwise it is impossible to express things like "track everything under this
particular directory". This would be important for luci-notify templates that
are in a separate directory.

Unfortunately, e.g. "*-dev.cfg" now means "*-dev.cfg files in the root config
set directory only". To allow the previous meaning of "*-dev.cfg files
regardless of the directory", we introduce new pattern syntax: "**/<pat>".

This is expected to be backward compatible since up till now all generated
configs were in a single directory.

R=tandrii@chromium.org
BUG=833946

Change-Id: I84d0c445e1306dc25c76899de5ddeb13c9d41b17