[bb] One client
Currently each piece of code that needs a HTTP or Buildbucket client, creates
one. Move that code into (*baseCommandRun).initClients, and save them
into the fields and call the new method in the beginning of a field.
Reuse same client for all requests.

Reuse created HTTP client for Gerrit RPCs as well.

Bug: 946422
Change-Id: Id7993e5f03f515c27c2fde1ef3465b0fbd6e69b8