Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Let me clarify your comment regarding "generalizing too much is not valuable". Do you think this applies to the CL? If so, I'm ok to abandon.

Yes, I think this CL is around net ~=0 in maintainability of this code (i.e. it doesn't make it worse, but doesn't improve it much either).

There are two conflicting concerns with this kind of code:
  1. Copy pastas is generally bad. Making changes requires updating all places.
  2. Excessive layering is also generally bad. Makes it hard to figure out what is going on.

I usually follow the rule: if copy pasta is *not 100% identical* and there are at most 2 instances of it, I keep it as is. When it becomes 3 or more, I refactor.

The reason is that fixing this situation (2 instances of slightly different copy pasta) requires adding "tweaks" to the "general" single implementation (in your CL this is Validator callback, and a flag to hide secret token in LogRequest). And I believe the complexity of this additional layer with tweaks outweights the benefit of reducing copy pastas of *2 instances only*. Once there are more instances of copy pasta, deduplication becomes more beneficial.

Either way, I don't feel too strongly about this. If you fix two comments I left, this CL is good to go.

Patch-set: 1
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>