[milo] Cut botlist at 1000 bots, show error message if over
Currently Milo is failing to store the entity because the entity is too large.
Cut the bot list to 1000 bots, and show a message if it is at 1000 to click on swarming.

The max size for a GCE instance name is 128 bytes, so the max size of the entity
at 1000 bots should be under 150KB, which is under the 1MB datastore limit.

Bug: 947601
Change-Id: I641dfc586abf4466904757117ec543ae4d5870cf