[bb] ls: require patchset
bb is currently being smart about unspecified patchset of a CL.
It is defaulted to the latest patchset.
In practice it is confusing when used in bb ls because it is different from
Gerrit which displays builds of a range of patchsets.

Also fix the bug of overwriting a user-specified patchset

Change-Id: I1aec923b50c2382c18f57b20c8ecf52ef5facffb