[GCE] Generate hostname when creating VM entity
Previously a VM entity was first written to the datastore without a
hostname, then on first fetch a hostname was assigned and written back
to the datastore. Once set, hostname never changes. Therefore, to save
an extra write, assign it on VM entity creation instead.

Bug: 897355
Change-Id: Ia0ab7ded3b27e711a2b437fb60bf7316e6d4ced9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1566444
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: smut <smut@google.com>