Limit the timeline group titles to 50ems.
Limit the timeline group titles to 50 ems to prevent the situation
where the group title is so wide that it does not allow for the
rendering of the body of the timeline widget.

This may be followed by further fix ups where we only show the last
text segement in a nested step or collapse nested steps under the
parent step. This change, however, will quickly get the timeline
usable again on laptops and seems necessary anyway for these new
style long step description text steps.

Bug: 953278
Change-Id: I6d70ea738a5d0ea7419f6c1cec69a144e57cc340