Make it possible to set "raw" format as default, but warn users about it
Many logs are loading several times faster when using ?format=raw, but adding
this suffix to the URL each time is tedious. Hence, we make it possible to raw
format as default, but warn users that going back requires manually removing
cookies as we are not able to add any links/javascript to non-HTML format.

R=hinoka@chromium.org

Bug: 940946
Change-Id: If650f8ebb6c27d17f916ed2c20f3d63ab139364d