[GCE] Work around rate limiting
GCE returns "Rate Limit Exceeded" without a known transient HTTP status
code, so do not delete the VM entity when this happens. Continue to
check the same hostname until GCE returns a definitive response.

Also decrease the throughput of the instance creation task queue to
try and avoid "Rate Limit Exceeded" in the first place, and decrease the
frequency of the cron job which launches instance creation task queues
to compensate.

Bug: 954632
Change-Id: Ief59eb5569cc5f55a201bb18f3fe07af028b3b03