[GCE] Don't err when the VM whose bot to manage doesn't exist
Management tasks are triggered every minute, and if necessary the
management task will trigger another task to delete the VM. Occasionally
this task is delayed and only completes after the next management task
has already been scheduled. This results in the second management task
seeing a non-existent VM entity, or a recreated VM entity with a
non-existent GCE instance. In such cases, the management task currently
errs, which is expected and correct behavior, but produces an error
message in the logs making it look unexpected or incorrect. For these
cases, don't err.

Bug: 897355
Change-Id: I276fabb0b96d1e265786334550ed0e1715c6b085
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1583091
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: smut <smut@google.com>