Update patch set 7
Patch Set 7:

because this isn't a public api, I think it's fine as-is; it's more important to unblock than to get this perfect in the first try. My original comments assumed that some user code might see this, or that overhauling this would be difficult.

Patch-set: 7
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>