Update patch set 2
Patch Set 2:

(2 comments)

Ok, this is closer to what I was expecting :D

Still have a question, but this is looking good to me.

It would be great if we could add a regression test for this as well (i.e. have a 'decoy' in path and a later match)

Patch-set: 2
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>